% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph.R
\name{hydrograph}
\alias{hydrograph}
\title{Hydrograph plus Hyetograph}
\usage{
hydrograph(
  data,
  time = time,
  runoff = Runoff,
  rain = Rain,
  agg.time_unit = "3 hours",
  agg.window = 8,
  vsep = 0.5
)
}
\arguments{
\item{data}{tbl or Data frame object.}

\item{time}{Vector name of timestamp variable in `data`.}

\item{runoff}{Vector name of Runoff variable in `data`.}

\item{rain}{Vector name of Rainfall variable in `data`.}

\item{agg.time_unit}{Integrer. Default "3 hours". Time lag unit to aggregate
rainfall bars by sum function. This will be showed as light blue line of
hyetrograph.}

\item{agg.window}{Integrer. Default 8. Value of number o values from rainfall
data to aggregate by rolling mean function. This will be showed as dark
blue line of hyetrograph.}

\item{vsep}{numeric. Default 0.5. Value between 0 and 1 to scale the vertical
separation of hyetograph from hydrograph. 1 means 100% of difference
between min and max runoff.}
}
\description{
The function to plot a hyetograph over a hydrograph based on ggplot functions.
}
\details{
ggplot2 object is generated
}
\note{
Based on the request at \url(https://stackoverflow.com/questions/42057832/how-to-draw-rainfall-runoff-graph-in-r-using-ggplot)
}
\examples{
# ploting soil moisture
data("zhurucay")
#
hydrograph(data = zhurucay,
    time = Fecha,
    runoff = `Caudal (m3/s)`,
    rain = `Lluvia (mm)`,
    agg.time_unit = "month",
    agg.window = 6
)
}
\author{
Gabriel Gaona
}
\keyword{hydrograph,}
\keyword{plot}
\keyword{plot,}
\keyword{rainfall-runoff}
