\name{hydroTools-package}
\alias{hydroTools-package}
\alias{hydroTools}
\docType{package}
\title{
\packageTitle{hydroTools}
}
\description{
\packageDescription{hydroTools}
}
\details{

\packageDESCRIPTION{hydroTools}
\packageIndices{hydroTools}

This packages provides a collection of tools and functions easing the access and handling of hydrological data. Functions inlcude reading time series from files into timeseries objects, as well as spatiala nd spatio-temporal data up to grouping and disaggregating functions commonly used.

}
\author{
\packageAuthor{hydroTools}

Maintainer: \packageMaintainer{hydroTools}
}
\keyword{ package }
