% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\arguments{
\item{lhs, rhs}{A chain process to concatenate}
}
\description{
Like dplyr, hydroutils also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of
ggplot(mtcars, aes(mpg, wt)) + geom_points()
# you can write
mtcars \%>\% ggplot(aes(mpg, wt)) + geom_points()
}
