% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnf_unit_convert.R
\name{rnf_unit_convert}
\alias{rnf_unit_convert}
\title{Runoff conversion Volumen per time from/to level}
\usage{
rnf_unit_convert(x, from = "vol", to.time.unit = "day", area)
}
\arguments{
\item{x}{Vector numeric. Runoff variable}

\item{from}{Character. String to convert 'vol' or 'lev'. Default 'vol'}

\item{to.time.unit}{Character. String containing the time units of conversion.
See more in details}

\item{area}{numeric, catchment area  in suare meters.}
}
\description{
The function to transform runoff variables from/to cubic meters
per second to water level in mm per unit time on a square meter.
}
\details{
Runoff unit conversion

Runoff units can be compared with precipitation by transforming it scale into
comparable scales. Usually rainfall is measured in level terms meaning the
rise of water level in a theoretical volume of base equal to one square meter.



`from` argument require `"vol"` meaning "volume per second" (m3/s) or `"lev"`
meaning "level on a time over area" (mm day).

`to.time.unit` is a time fraction string i.e. "day", "3 hour", "10 minutes", etc.
}
\examples{

# In a catchment area of 1.4 km2 a measure of water outcomes by runoff is
# 0.023 'm3/s'
area <- 1.4 * 1000000 #convert km2 to m2
r_vol <- 0.023
# let's convert to "mm dia"
r_lev <- rnf_unit_convert(r_vol, from = "vol",
                          to.time.unit = "1 day", area = area)
# Reverse process
r_lev <- rnf_unit_convert(r_lev, from = "lev",
                          to.time.unit = "1 day", area = area)
}
\author{
Gabriel Gaona
}
\keyword{conversion,}
\keyword{plot}
\keyword{rainfall-runoff,}
