% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_moisture.R
\docType{data}
\name{soil_moisture}
\alias{soil_moisture}
\title{Soil moisture data for examples.}
\format{A tibble with 16326 rows and 3 variables:
\describe{
  \item{timestamp}{Timestamp with breaks each 5 min}
  \item{runoff}{Fraction of water content per volume of soil}
  \item{rain}{Total rainfall in 5 min}

}}
\source{
\url{http://www.tmf_respect.org/}
}
\usage{
soil_moisture
}
\description{
A dataset containing the data needed for a hydrograph with hyetograph. Two
    month of 5 min records of soil moisture and rainfall data
}
\author{
Gabriel Gaona
}
\keyword{datasets}
